Move config out
